pago46/requirements.txt
Django
psycopg2-binary
djangorestframework
graphene-django
pytest
model_bakery
pytest-django
freezegun